chore(svm): extract seed derivation from user input to util
